Allow creation of local variables & enable editing on restricted instances

+++ api/apiv1/templatevars.go
@@ -7,6 +7,7 @@ import (
 	"github.com/paularlott/knot/database"
 	"github.com/paularlott/knot/database/model"
 	"github.com/paularlott/knot/internal/origin_leaf/leaf"
+	"github.com/paularlott/knot/internal/origin_leaf/origin"
 	"github.com/paularlott/knot/util/rest"
 	"github.com/paularlott/knot/util/validate"
 
@@ -15,7 +16,7 @@ import (
 
 func HandleGetTemplateVars(w http.ResponseWriter, r *http.Request) {
 	remoteClient := r.Context().Value("remote_client")
-	if remoteClient != nil {
+	if !origin.RestrictedLeaf && remoteClient != nil {
 		client := remoteClient.(*apiclient.ApiClient)
 		templateVars, code, err := client.GetTemplateVars()
 		if err != nil {
@@ -38,10 +39,15 @@ func HandleGetTemplateVars(w http.ResponseWriter, r *http.Request) {
 		}
 
 		for _, variable := range templateVars {
+			if origin.RestrictedLeaf && !variable.Local {
+				continue
+			}
+
 			v := apiclient.TemplateVar{
 				Id:        variable.Id,
 				Name:      variable.Name,
 				Location:  variable.Location,
+				Local:     variable.Local,
 				Protected: variable.Protected,
 			}
 			data.TemplateVar = append(data.TemplateVar, v)
@@ -76,9 +82,9 @@ func HandleUpdateTemplateVar(w http.ResponseWriter, r *http.Request) {
 	}
 
 	remoteClient := r.Context().Value("remote_client")
-	if remoteClient != nil {
+	if !origin.RestrictedLeaf && remoteClient != nil {
 		client := remoteClient.(*apiclient.ApiClient)
-		code, err := client.UpdateTemplateVar(templateVarId, request.Name, request.Location, request.Value, request.Protected)
+		code, err := client.UpdateTemplateVar(templateVarId, request.Name, request.Location, request.Local, request.Value, request.Protected)
 		if err != nil {
 			rest.SendJSON(code, w, ErrorResponse{Error: err.Error()})
 			return
@@ -93,8 +99,14 @@ func HandleUpdateTemplateVar(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
+		if origin.RestrictedLeaf && !templateVar.Local {
+			rest.SendJSON(http.StatusForbidden, w, ErrorResponse{Error: "No permission to manage template variables"})
+			return
+		}
+
 		templateVar.Name = request.Name
 		templateVar.Location = request.Location
+		templateVar.Local = request.Local
 		templateVar.Value = request.Value
 		templateVar.Protected = request.Protected
 		templateVar.UpdatedUserId = user.Id
@@ -137,19 +149,23 @@ func HandleCreateTemplateVar(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	if origin.RestrictedLeaf && !request.Local {
+		request.Local = true
+	}
+
 	remoteClient := r.Context().Value("remote_client")
-	if remoteClient != nil {
+	if !origin.RestrictedLeaf && remoteClient != nil {
 		var code int
 		var err error
 
 		client := remoteClient.(*apiclient.ApiClient)
-		id, code, err = client.CreateTemplateVar(request.Name, request.Location, request.Value, request.Protected)
+		id, code, err = client.CreateTemplateVar(request.Name, request.Location, request.Local, request.Value, request.Protected)
 		if err != nil {
 			rest.SendJSON(code, w, ErrorResponse{Error: err.Error()})
 			return
 		}
 	} else {
-		templateVar := model.NewTemplateVar(request.Name, request.Location, request.Value, request.Protected, user.Id)
+		templateVar := model.NewTemplateVar(request.Name, request.Location, request.Local, request.Value, request.Protected, user.Id)
 
 		err = db.SaveTemplateVar(templateVar)
 		if err != nil {
@@ -173,7 +189,7 @@ func HandleDeleteTemplateVar(w http.ResponseWriter, r *http.Request) {
 	templateVarId := chi.URLParam(r, "templatevar_id")
 
 	remoteClient := r.Context().Value("remote_client")
-	if remoteClient != nil {
+	if !origin.RestrictedLeaf && remoteClient != nil {
 		client := remoteClient.(*apiclient.ApiClient)
 		code, err := client.DeleteTemplateVar(templateVarId)
 		if err != nil {
@@ -185,6 +201,11 @@ func HandleDeleteTemplateVar(w http.ResponseWriter, r *http.Request) {
 		templateVar, err := db.GetTemplateVar(templateVarId)
 		if err != nil {
 			rest.SendJSON(http.StatusInternalServerError, w, ErrorResponse{Error: err.Error()})
+			return
+		}
+
+		if origin.RestrictedLeaf && !templateVar.Local {
+			rest.SendJSON(http.StatusForbidden, w, ErrorResponse{Error: "No permission to manage template variables"})
 			return
 		}
 
@@ -205,7 +226,7 @@ func HandleGetTemplateVar(w http.ResponseWriter, r *http.Request) {
 	templateVarId := chi.URLParam(r, "templatevar_id")
 
 	remoteClient := r.Context().Value("remote_client")
-	if remoteClient != nil {
+	if !origin.RestrictedLeaf && remoteClient != nil {
 		client := remoteClient.(*apiclient.ApiClient)
 		templateVar, code, err := client.GetTemplateVar(templateVarId)
 		if err != nil {
@@ -221,7 +242,7 @@ func HandleGetTemplateVar(w http.ResponseWriter, r *http.Request) {
 			rest.SendJSON(http.StatusNotFound, w, ErrorResponse{Error: err.Error()})
 			return
 		}
-		if templateVar == nil {
+		if templateVar == nil || (origin.RestrictedLeaf && !templateVar.Local) {
 			rest.SendJSON(http.StatusNotFound, w, ErrorResponse{Error: "Template variable not found"})
 			return
 		}
@@ -238,6 +259,7 @@ func HandleGetTemplateVar(w http.ResponseWriter, r *http.Request) {
 			Name:      templateVar.Name,
 			Value:     val,
 			Location:  templateVar.Location,
+			Local:     templateVar.Local,
 			Protected: templateVar.Protected,
 		}
